Reduce priority of input_text_display_with_cursor modifier
This may conflict with go.py if it has an equal or higher priority than
that, which causes go.py to not show any buffers. So set it to a low
priority so it doesn't conflict with anything else. Since this only
prepends to the input string and doesn't read the string, I think it
could have the lowest priority, but leave some room for lower priorities
just in case another hook needs it.