Unescape from_url in attachments
It looks like from_url should be unescaped, but title_link and image_url
should not. Slack seems a bit inconsistent here though, as an image in
an attachment will use the unescaped title_link, but the link in the
title itself will use the original (possibly escaped) one. And the title
for the image link will use the original (possibly escaped) title, while
the title in the attachment will unescape it.

I suppose you're not supposed to HTML escape (or Slacks version of it)
URLs in title_link and image_url (which makes sense), but Slack does
escape from_url if you send a message with a link to a thread message,
so we have to handle that.