Don't print the same message multiple times
When joining a channel you seem to get the last messages for it over the
WebSocket. When joining a channel you've parted and you still have the
buffer for open, those messages will already be in the buffer, so we
have to make sure their not printed multiple times.

I've also been told that messages can appear multiple times in the
reconnect brach, which this should fix too.