Handle errors when fetching nick
There's an error with a bot for a certain user on #wee-slack where the
bots user id is set to an item_not_found error in users when it tries to
print the nick. Since the bot has a user id, it's requested along with
the other users when initialize_items is called. However for some reason
I haven't figured out it's not included in the response. When the user
tries to request the user id directly with curl it works, so I don't
know why it's missing in the initialize_items request.

To workaround this, prefer looking up the bot id before user id and
handle errors when requesting either. In commit
d094323c212587c8bdb68fda7dac2f053143ad32 the nick resolving was changed
to prefer user id before bot id because requesting the bot id of
Slackbot fails. By handling the error we get from that it will fall back
to looking up the user id when the bot id request fails.